bhiv-core-bucket-karma-prana-orchrstration-insightcore-insightflow/v1-BHIV_CORE-main/requirements.txt
fastapi
uvicorn
requests
requests-toolbelt
pydantic
motor
pyPDF2
pdfplumber
python-dotenv
qdrant-client
langchain-core
langchain-huggingface
langchain-community
sentence-transformers
transformers
speechrecognition
pillow
pytest
pytest-asyncio
python-multipart
torch
torchaudio
soundfile
librosa
pydub
speechrecognition
pymongo
nltk
spacy
tqdm
jinja2
python-jose[cryptography]
passlib[bcrypt]
passlib
pdfplumber
qdrant-client 
sentence-transformers
qdrant-client[fastembed]